requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
numpy
pandas==1.3.5
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.28.0
six==1.16.0
SQLAlchemy==1.3.24
urllib3==1.26.9